Refactor to make draw rect generic
